Add parameter "dump_always".
The default behaviour is to only dump the info page contents to the
console if they are not valid. This is usually enough as the
contents are logged as an event.
Without Kafka support like in plain CLI tests, the data is lost. The
new parameter "dump_always" forces a dump on the console to fix this.